Rename `syscon::clocksource` to `clock_source`
This follows the typical Rust naming style better.
