+++ examples/usart.rs
@@ -4,7 +4,7 @@
 extern crate panic_halt;
 
 use lpc8xx_hal::{
-    cortex_m_rt::entry, prelude::*, syscon::clocksource::UsartClock,
+    cortex_m_rt::entry, prelude::*, syscon::clock_source::UsartClock,
     Peripherals,
 };
